Refacto queryResultTypeWalkerTest
